Separate memory dumping utilities from MemoryManager
